fix: fixup intervals for ban parameters
This is completely my bad and an oversight from the lack of tf docs -> https://registry.terraform.io/providers/hashicorp/google/latest/docs/resources/compute_security_policy#nested_threshold
Checking GCP -> https://cloud.google.com/armor/docs/rate-limiting-overview#ban-clients
```
ban_duration_sec: This is the additional number of seconds for which a client is banned after the interval_sec period elapses. The value must be 60, 120, 180, 240, 300, 600, 900, 1200, 1800, 2700, or 3600 seconds.

interval_sec: The number of seconds in the time interval. The value must be 10, 30, 60, 120, 180, 240, 300, 600, 900, 1200, 1800, 2700, or 3600 seconds.
```